Move unify methods for dynamics in `dynamics.py`
- Deprecate wn.simulators.infrastructure.
- Update documentation, example notebooks, and unit tests.
