+++ whynot/__init__.py
@@ -3,13 +3,14 @@
 __version__ = "0.10.14"
 from whynot.algorithms import *
 from whynot.simulators import *
-from whynot import causal_graphs
-from whynot import framework
+from whynot import causal_graphs, dynamics, framework
+from whynot.dynamics import (
+    DynamicsExperiment,
+    Run,
+)
 from whynot.framework import (
     Dataset,
-    DynamicsExperiment,
     InferenceResult,
     parameter,
-    Run,
 )
 from whynot import utils
