+++ whynot/causal_graphs.py
@@ -120,7 +120,7 @@ def run_to_box(run, trace_ctx, skip_args=None):
 
     Parameters
     ----------
-        run: whynot.framework.Run
+        run: whynot.dynamics.Run
             The run for the user to trace.
         trace_ctx:  trace_stack context manager
             Every box is created within the context of this trace
@@ -130,7 +130,7 @@ def run_to_box(run, trace_ctx, skip_args=None):
 
     Returns
     -------
-        boxed_run: whynot.framework.Run
+        boxed_run: whynot.dynamics.Run
             Run object where each State has all of its fields (except those
             listed in skip_args) replaced with a Box object for tracing.
 
@@ -149,7 +149,7 @@ def run_to_box(run, trace_ctx, skip_args=None):
         )
         boxed_states.append(boxed_state)
         global_nodemap.update(nodemap)
-    boxed_run = wn.framework.Run(states=boxed_states, times=run.times)
+    boxed_run = wn.dynamics.Run(states=boxed_states, times=run.times)
     return boxed_run, global_nodemap
 
 
@@ -427,8 +427,8 @@ def build_dynamics_graph(simulator, runs, config, config_nodes=False):
         simulator: whynot simulator module
             Simulator dynamics to trace, e.g. whynot.hiv.
         runs: list
-            List of whynot.framework.Runs objects generated by the simulator
-        config: whynot.simulators.infrastructure.BaseConfig
+            List of whynot.dynamics.Runs objects generated by the simulator
+        config: whynot.dynamics.BaseConfig
             Config object use for all of the simulator runs
         config_nodes: bool
             Whether or not to add nodes and edges to the graphs for
@@ -513,12 +513,12 @@ def ate_graph_builder(
     ----------
         simulator: whynot.simulator module
             The simulator used to conduct the experiment.
-        run: whynot.framework.Run
+        run: whynot.dynamics.Run
             A single run object generated from the simulator under the config
             and intervention.
-        config: whynot.simulators.infrastructure.BaseConfig
+        config: whynot.dynamics.BaseConfig
             Configuration object used for the experiment.
-        intervention: whynot.simulators.infrastructure.BaseIntervention
+        intervention: whynot.dynamics.BaseIntervention
             Intervention object used for the experiment.
         treatment_dependencies: list
             List of node names treatment depends on.
