+++ whynot/simulators/lotka_volterra/simulator.py
@@ -5,7 +5,7 @@ import numpy as np
 from scipy.integrate import odeint
 
 import whynot as wn
-from whynot.simulators.infrastructure import BaseConfig, BaseIntervention, BaseState
+from whynot.dynamics import BaseConfig, BaseIntervention, BaseState
 
 
 @dataclasses.dataclass
@@ -134,7 +134,7 @@ def simulate(initial_state, config, intervention=None, seed=None):
 
     Returns
     -------
-        run: whynot.framework.Run
+        run: whynot.dynamics.Run
             Simulator rollout
 
     """
@@ -153,4 +153,4 @@ def simulate(initial_state, config, intervention=None, seed=None):
     )
 
     states = [initial_state] + [State(*state) for state in solution[1:]]
-    return wn.framework.Run(states=states, times=t_eval)
+    return wn.dynamics.Run(states=states, times=t_eval)
